nextest-runner: fix performance regression
Fix performance regression introduced in #97 by spawing a task onto a
threadpool to wait for the test to complete, signaling back to the main
thread that its completed via a channel. The main thread then waits on
recieving this message using a timeout, logging each time a timeout is
reached.